Initial fix for XTENLANG-530:   For every static field f the compiler generates     an accessor method, f__get()     a volatile status field, f__status     a local initializer, f__do_init()     a global initializer, f__init() (which invokes f__do_init() in place 0)     a broadcast id f__id     a broadcast deserializer function f__deserialize().   Fields are initialized on first access.   Fields use "when" to synchronize initialization     so a circular initialization dependency results in deadlock.   Static initializer blocks are disabled in codegen, but not yet in the parser.   Removed _static_init methods from all classes.   initialize_xrx() now initializes the necessary fields individually.   LIMITATION: does not work for struct fields.
Async deserializers can now return null (which will cause the closure to be ignored).

Plus minor tweaks of codegen and C++ implementation.


git-svn-id: svn://svn.code.sf.net/p/x10/code/trunk@11325 a326200e-df2c-42e0-81fa-597e909af41d

+++ x10.runtime/src-cpp/x10aux/deserialization_dispatcher.cc
@@ -2,6 +2,7 @@
 #include <x10aux/deserialization_dispatcher.h>
 #include <x10aux/serialization.h>
 #include <x10aux/network.h>
+#include <x10aux/alloc.h>
 
 #include <assert.h>
 
@@ -100,7 +101,7 @@ void DeserializationDispatcher::registerHandlers () {
 }
 
 void DeserializationDispatcher::registerHandlers_ () {
-    for (int i=0 ; i<next_id ; ++i) {
+    for (size_t i=0 ; i<next_id ; ++i) {
         if (i<deser_sz && deser_v[i]) {
             _S_("(DeserializationDispatcher registered id "<<i<<" as an async)");
             x10aux::register_async_handler(i);
